move dto/db models to correct data package
