+++ app/src/main/java/com/example/tasky/feature_login/domain/repository/UserRemoteRepository.kt
@@ -2,9 +2,9 @@ package com.example.tasky.feature_login.domain.repository
 
 import com.example.tasky.common.data.util.DataError
 import com.example.tasky.common.data.util.Result
-import com.example.tasky.feature_login.domain.model.LoginUserInfo
-import com.example.tasky.feature_login.domain.model.LoginUserResponse
-import com.example.tasky.feature_login.domain.model.RegisterUserInfo
+import com.example.tasky.feature_login.data.model.LoginUserInfo
+import com.example.tasky.feature_login.data.model.LoginUserResponse
+import com.example.tasky.feature_login.data.model.RegisterUserInfo
 
 interface UserRemoteRepository {
     suspend fun registerUser(registerUserInfo: RegisterUserInfo): Result<Unit, DataError.Network>
